net.c: Fixed some type casting warnings from splint.

+++ net.c
@@ -21,8 +21,8 @@ static Event_move
 mk_event_move (Byte * d){
   Event_move e;
   e.t    = EVENT_MOVE;
-  e.u    = d[1];
-  e.dir  = d[2];
+  e.u    = (int)d[1];
+  e.dir  = (int)d[2];
   return(e);
 }
 
@@ -32,9 +32,9 @@ static Event_range
 mk_event_range (Byte * d){
   Event_range e;
   e.t    = EVENT_RANGE;
-  e.a    = d[1];
-  e.d    = d[2];
-  e.dmg  = d[3];
+  e.a    = (int)d[1];
+  e.d    = (int)d[2];
+  e.dmg  = (int)d[3];
   return(e);
 }
 
@@ -44,10 +44,10 @@ static Event_melee
 mk_event_melee (Byte * d){
   Event_melee e;
   e.t                = EVENT_MELEE;
-  e.a                = d[1];
-  e.d                = d[2];
-  e.attackers_killed = d[3];
-  e.defenders_killed = d[4];
+  e.a                = (int)d[1];
+  e.d                = (int)d[2];
+  e.attackers_killed = (int)d[3];
+  e.defenders_killed = (int)d[4];
   return(e);
 }
 
@@ -57,8 +57,8 @@ static Event_endturn
 mk_event_endturn (Byte * d){
   Event_endturn e;
   e.t          = EVENT_ENDTURN;
-  e.old_player = d[1];
-  e.new_player = d[2];
+  e.old_player = (int)d[1];
+  e.new_player = (int)d[2];
   return(e);
 }
 
@@ -68,7 +68,7 @@ static Event_death
 mk_event_death (Byte *d){
   Event_death e;
   e.t = EVENT_DEATH;
-  e.u = *id2unit(d[1]);
+  e.u = *id2unit((int)d[1]);
   return(e);
 }
 
@@ -79,10 +79,10 @@ send_move (Event_move e){
   Byte size = 3;
   Byte d[3];
   SDLNet_TCP_Send(socket, &size, 1);
-  d[0] = e.t;
-  d[1] = e.u;
-  d[2] = e.dir;
-  SDLNet_TCP_Send(socket, d, size);
+  d[0] = (Byte)e.t;
+  d[1] = (Byte)e.u;
+  d[2] = (Byte)e.dir;
+  SDLNet_TCP_Send(socket, d, (int)size);
 }
 
 
@@ -92,12 +92,12 @@ send_melee (Event_melee e){
   Byte size = 5;
   Byte d[5];
   SDLNet_TCP_Send(socket, &size, 1);
-  d[0] = e.t;
-  d[1] = e.a;
-  d[2] = e.d;
-  d[3] = e.attackers_killed;
-  d[4] = e.defenders_killed;
-  SDLNet_TCP_Send(socket, d, size);
+  d[0] = (Byte)e.t;
+  d[1] = (Byte)e.a;
+  d[2] = (Byte)e.d;
+  d[3] = (Byte)e.attackers_killed;
+  d[4] = (Byte)e.defenders_killed;
+  SDLNet_TCP_Send(socket, d, (int)size);
 }
 
 
@@ -107,11 +107,11 @@ send_range (Event_range e){
   Byte size = 4;
   Byte d[4];
   SDLNet_TCP_Send(socket, &size, 1);
-  d[0] = e.t;
-  d[1] = e.a;
-  d[2] = e.d;
-  d[3] = e.dmg;
-  SDLNet_TCP_Send(socket, d, size);
+  d[0] = (Byte)e.t;
+  d[1] = (Byte)e.a;
+  d[2] = (Byte)e.d;
+  d[3] = (Byte)e.dmg;
+  SDLNet_TCP_Send(socket, d, (int)size);
 }
 
 
@@ -122,10 +122,10 @@ send_endturn (Event_endturn e){
   Byte size = 3;
   Byte d[3];
   SDLNet_TCP_Send(socket, &size, 1);
-  d[0] = e.t;
-  d[1] = e.old_player;
-  d[2] = e.new_player;
-  SDLNet_TCP_Send(socket, d, size);
+  d[0] = (Byte)e.t;
+  d[1] = (Byte)e.old_player;
+  d[2] = (Byte)e.new_player;
+  SDLNet_TCP_Send(socket, d, (int)size);
 }
 
 
@@ -135,9 +135,9 @@ send_death (Event_death e){
   Byte size = 2;
   Byte d[2];
   SDLNet_TCP_Send(socket, &size, 1);
-  d[0] = e.t;
-  d[1] = e.u.id;
-  SDLNet_TCP_Send(socket, d, size);
+  d[0] = (Byte)e.t;
+  d[1] = (Byte)e.u.id;
+  SDLNet_TCP_Send(socket, d, (int)size);
 }
 
 
@@ -146,10 +146,10 @@ static void
 get_data (Byte * data, Byte * size){
   int bytes_recieved;
   SDLNet_TCP_Recv(socket, size, 1);
-  bytes_recieved = SDLNet_TCP_Recv(socket, data, *size);
-  if(bytes_recieved != *size){
+  bytes_recieved = SDLNet_TCP_Recv(socket, data, (int)*size);
+  if(bytes_recieved != (int)*size){
     printf("recieved: %i, expected: %i", 
-        bytes_recieved, (unsigned int)size);
+        bytes_recieved, (int)size);
   }
 }
 
@@ -183,7 +183,7 @@ send_event (Event e) {
 
 void
 send_int_as_uint8 (int n){
-  Byte x = n;
+  Byte x = (Byte)n;
   SDLNet_TCP_Send(socket, &x, 1);
 }
 
@@ -222,7 +222,7 @@ void
 init_network (char * hostname, int port){
   IPaddress ip;
   SDLNet_Init();
-  SDLNet_ResolveHost(&ip, hostname, port);
+  SDLNet_ResolveHost(&ip, hostname, (Uint16)port);
   socket = SDLNet_TCP_Open(&ip);
   if(!socket) {
     puts("cannot open socket"); 
@@ -243,7 +243,7 @@ get_scenario_name_from_server (char * str){
       /*incoming data size in bytes*/
       Byte size;
       SDLNet_TCP_Recv(socket, &size, 1);
-      SDLNet_TCP_Recv(socket, str, size);
+      SDLNet_TCP_Recv(socket, str, (int)size);
       return;
     }
   }
